refactor: create component not-found
