--- src/app/features/error/pages/not-found/not-found.component.html
@@ -1,42 +1,0 @@
-<app-header></app-header>
-
-<app-breadcrumb></app-breadcrumb>
-
-<app-layout-content containerMarginTop="mt-0" containerMarginBottom="mb-10">
-  <div class="w-full max-w-[500px]">
-    <p-image src="images/not-found.svg" alt="page not found" />
-
-    <div class="w-full flex-col items-center">
-      <h5
-        class="text-3xl text-gray-700 leading-8 tracking-[0.03rem] font-poppins font-semibold mb-9 text-center"
-      >
-        404, Page not found.
-      </h5>
-      <p
-        class="text-sm text-gray-500 leading-6 tracking-[0.03rem] font-poppins font-light text-center"
-      >
-        Something went wrong. It’s look that your requested could not be found.
-        It’s look like the link is broken or the page is removed.
-      </p>
-
-      <div
-        class="flex flex-col md:flex-row items-center justify-center gap-4 mt-6"
-      >
-        <app-button-cta-large
-          type="submit"
-          label="Go back"
-          icon="pi pi-arrow-left"
-          (click)="goBack()"
-        ></app-button-cta-large>
-
-        <app-button-cta-large
-          type="submit"
-          label="Go to home"
-          icon="pi pi-home"
-          variant="outlined"
-          [routerLink]="['/' + routeEnum.HOME]"
-        ></app-button-cta-large>
-      </div>
-    </div>
-  </div>
-</app-layout-content>
