--- src/app/features/error/pages/not-found/not-found.component.ts
@@ -1,32 +1,0 @@
-import { Location } from '@angular/common';
-import { Component, inject } from '@angular/core';
-import { Router, RouterModule } from '@angular/router';
-import { RouteEnum } from '@enum/route.enum';
-import { BreadcrumbComponent } from '@shared/components/ui/breadcrumb/breadcrumb.component';
-import { ButtonCtaLargeComponent } from '@shared/components/ui/buttons/button-cta-large/button-cta-large.component';
-import { HeaderComponent } from '@shared/components/ui/header/header/header.component';
-import { LayoutContentComponent } from '@shared/components/ui/layout/layout-content/layout-content.component';
-import { ImageModule } from 'primeng/image';
-
-@Component({
-	selector: 'app-not-found',
-	imports: [
-		RouterModule,
-		HeaderComponent,
-		BreadcrumbComponent,
-		LayoutContentComponent,
-		ImageModule,
-		ButtonCtaLargeComponent,
-	],
-	templateUrl: './not-found.component.html',
-	styleUrl: './not-found.component.scss',
-})
-export class NotFoundComponent {
-	private _location: Location = inject(Location);
-
-	public routeEnum = RouteEnum;
-
-	goBack(): void {
-		this._location.back();
-	}
-}
