--- src/app/features/error/pages/product-not-found/product-not-found.component.html
@@ -1,42 +1,0 @@
-<app-header></app-header>
-
-<app-breadcrumb></app-breadcrumb>
-
-<app-layout-content containerMarginTop="mt-0" containerMarginBottom="mb-10">
-  <div class="w-full max-w-[500px]">
-    <p-image src="images/not-found.svg" alt="page not found" />
-
-    <div class="w-full flex-col items-center">
-      <h5
-        class="text-3xl text-gray-700 leading-8 tracking-[0.03rem] font-poppins font-semibold mb-9 text-center"
-      >
-        404, Product not found.
-      </h5>
-      <p
-        class="text-sm text-gray-500 leading-6 tracking-[0.03rem] font-poppins font-light text-center"
-      >
-        Oops! It looks like the product you're trying to view no longer exists
-        or has been moved. You can browse our catalog to find a similar product.
-      </p>
-
-      <div
-        class="flex flex-col md:flex-row items-center justify-center gap-4 mt-6"
-      >
-        <app-button-cta-large
-          type="submit"
-          label="Go back"
-          icon="pi pi-arrow-left"
-          (click)="goBack()"
-        ></app-button-cta-large>
-
-        <app-button-cta-large
-          type="submit"
-          label="Go to products"
-          icon="pi pi-home"
-          variant="outlined"
-          [routerLink]="['/' + routeEnum.PRODUCT]"
-        ></app-button-cta-large>
-      </div>
-    </div>
-  </div>
-</app-layout-content>
