+++ src/app/routes/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 import { RouteEnum } from '@enum/route.enum';
 import { AboutPageComponent } from '@features/about/pages/about-page/about-page.component';
-import { NotFoundComponent } from '@features/error/pages/not-found/not-found.component';
+import { NotFoundPageComponent } from '@features/error/pages/not-found-page/not-found-page.component';
 import { FaqPageComponent } from '@features/faq/pages/faq-page/faq-page.component';
 import { HomePageComponent } from '@features/home/pages/home-page/home-page.component';
 
@@ -13,5 +13,5 @@ export const routes: Routes = [
 	{ path: RouteEnum.AUTH, loadChildren: () => import('./auth.routes').then(m => m.routes) },
 	{ path: RouteEnum.ACCOUNT, loadChildren: () => import('./account.routes').then(m => m.routes) },
 	{ path: RouteEnum.PRODUCT, loadChildren: () => import('./product.routes').then(m => m.routes) },
-	{ path: '**', component: NotFoundComponent, data: { title: '404 - Not found' } },
+	{ path: '**', component: NotFoundPageComponent, data: { title: '404 - Not found' } },
 ];
