+++ src/app/routes/product.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
 import { RouteEnum } from '@enum/route.enum';
-import { ProductNotFoundComponent } from '@features/error/pages/product-not-found/product-not-found.component';
+import { ProductNotFoundPageComponent } from '@features/error/pages/product-not-found-page/product-not-found-page.component';
 import { ProductsDetailsPageComponent } from '@features/product/products-details-page/products-details-page.component';
 import { ProductsPageComponent } from '@features/product/products-page/products-page.component';
 
@@ -8,7 +8,7 @@ export const routes: Routes = [
 	{ path: '', component: ProductsPageComponent, data: { title: 'products' } },
 	{
 		path: RouteEnum.NOT_FOUND,
-		component: ProductNotFoundComponent,
+		component: ProductNotFoundPageComponent,
 		data: { title: 'product not found' },
 	},
 	{
