+++ src/app/shared/components/ui/error/not-found/not-found.component.html
@@ -1,0 +1,33 @@
+<p-image [src]="imgSrc()" [alt]="imgAlt()" />
+
+<div class="w-full flex-col items-center">
+  <h5
+    class="text-3xl text-gray-700 leading-8 tracking-[0.03rem] font-poppins font-semibold mb-9 text-center"
+  >
+    {{ title() }}
+  </h5>
+  <p
+    class="text-sm text-gray-500 leading-6 tracking-[0.03rem] font-poppins font-light text-center"
+  >
+    {{ content() }}
+  </p>
+
+  <div class="flex flex-col md:flex-row items-center justify-center gap-4 mt-6">
+    @if (showBtnGoBack()) {
+    <app-button-cta-large
+      type="submit"
+      label="Go back"
+      icon="pi pi-arrow-left"
+      (click)="goBack()"
+    ></app-button-cta-large>
+    }
+
+    <app-button-cta-large
+      type="submit"
+      [label]="labelPrimaryLink()"
+      icon="pi pi-home"
+      variant="outlined"
+      [routerLink]="['/' + primaryLink()]"
+    ></app-button-cta-large>
+  </div>
+</div>
